wgml2: portability fixes

+++ bld/wgml2/c/gsfmin.c
@@ -59,7 +59,7 @@ condcode    scr_min( parm parms[MAX_FUN_PARMS], size_t parmcount, char * * resul
     int                 k;
     int                 len;
     getnum_block        gn;
-    long                minimum;
+    int                 minimum;
 
     (void)ressize;
 
@@ -68,7 +68,7 @@ condcode    scr_min( parm parms[MAX_FUN_PARMS], size_t parmcount, char * * resul
         return( cc );
     }
 
-    minimum = LONG_MAX;
+    minimum = INT_MAX;
 
     gn.ignore_blanks = false;
